requirements.txt
fastapi
uvicorn
pydantic
azure-identity
azure-devops
openai
python-dotenv
httpx
azure-cosmos
